Restructuring of TrafficAssignment
